Move ffi code, add kernel expr -> engine expr
FFI code for schema and expressions have been moved to their own
modules.

Add prototype of the visitor to enable kernel expresison ->
engine expression
