+++ ffi/src/lib.rs
@@ -10,8 +10,6 @@ use std::sync::Arc;
 use tracing::debug;
 use url::Url;
 
-use delta_kernel::expressions::{BinaryOperator, Expression, Scalar, UnaryOperator};
-use delta_kernel::schema::{ArrayType, DataType, MapType, PrimitiveType, StructType};
 use delta_kernel::snapshot::Snapshot;
 use delta_kernel::{DeltaResult, Engine, EngineData, Error, Table};
 use delta_kernel_ffi_macros::handle_descriptor;
@@ -32,7 +30,9 @@ use handle::Handle;
 extern crate self as delta_kernel_ffi;
 
 pub mod engine_funcs;
-pub mod scan;
+pub(crate) mod expressions;
+pub(crate) mod scan;
+pub(crate) mod schema;
 
 pub(crate) type NullableCvoid = Option<NonNull<c_void>>;
 
@@ -790,212 +790,6 @@ pub unsafe extern "C" fn free_string_slice_data(data: Handle<StringSliceIterator
     data.drop_handle();
 }
 
-/// The `EngineSchemaVisitor` defines a visitor system to allow engines to build their own
-/// representation of a schema from a particular schema within kernel.
-///
-/// The model is list based. When the kernel needs a list, it will ask engine to allocate one of a
-/// particular size. Once allocated the engine returns an `id`, which can be any integer identifier
-/// ([`usize`]) the engine wants, and will be passed back to the engine to identify the list in the
-/// future.
-///
-/// Every schema element the kernel visits belongs to some list of "sibling" elements. The schema
-/// itself is a list of schema elements, and every complex type (struct, map, array) contains a list
-/// of "child" elements.
-///  1. Before visiting schema or any complex type, the kernel asks the engine to allocate a list to
-///     hold its children
-///  2. When visiting any schema element, the kernel passes its parent's "child list" as the
-///     "sibling list" the element should be appended to:
-///      - For the top-level schema, visit each top-level column, passing the column's name and type
-///      - For a struct, first visit each struct field, passing the field's name, type, nullability,
-///        and metadata
-///      - For a map, visit the key and value, passing its special name ("map_key" or "map_value"),
-///        type, and value nullability (keys are never nullable)
-///      - For a list, visit the element, passing its special name ("array_element"), type, and
-///        nullability
-///  3. When visiting a complex schema element, the kernel also passes the "child list" containing
-///     that element's (already-visited) children.
-///  4. The [`visit_schema`] method returns the id of the list of top-level columns
-// WARNING: the visitor MUST NOT retain internal references to the string slices passed to visitor methods
-// TODO: struct nullability and field metadata
-#[repr(C)]
-pub struct EngineSchemaVisitor {
-    /// opaque state pointer
-    pub data: *mut c_void,
-    /// Creates a new field list, optionally reserving capacity up front
-    pub make_field_list: extern "C" fn(data: *mut c_void, reserve: usize) -> usize,
-
-    // visitor methods that should instantiate and append the appropriate type to the field list
-    /// Indicate that the schema contains a `Struct` type. The top level of a Schema is always a
-    /// `Struct`. The fields of the `Struct` are in the list identified by `child_list_id`.
-    pub visit_struct: extern "C" fn(
-        data: *mut c_void,
-        sibling_list_id: usize,
-        name: KernelStringSlice,
-        child_list_id: usize,
-    ),
-
-    /// Indicate that the schema contains an Array type. `child_list_id` will be a _one_ item list
-    /// with the array's element type
-    pub visit_array: extern "C" fn(
-        data: *mut c_void,
-        sibling_list_id: usize,
-        name: KernelStringSlice,
-        contains_null: bool, // if this array can contain null values
-        child_list_id: usize,
-    ),
-
-    /// Indicate that the schema contains an Map type. `child_list_id` will be a _two_ item list
-    /// where the first element is the map's key type and the second element is the
-    /// map's value type
-    pub visit_map: extern "C" fn(
-        data: *mut c_void,
-        sibling_list_id: usize,
-        name: KernelStringSlice,
-        value_contains_null: bool, // if this map can contain null values
-        child_list_id: usize,
-    ),
-
-    /// visit a `decimal` with the specified `precision` and `scale`
-    pub visit_decimal: extern "C" fn(
-        data: *mut c_void,
-        sibling_list_id: usize,
-        name: KernelStringSlice,
-        precision: u8,
-        scale: u8,
-    ),
-
-    /// Visit a `string` belonging to the list identified by `sibling_list_id`.
-    pub visit_string:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `long` belonging to the list identified by `sibling_list_id`.
-    pub visit_long:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit an `integer` belonging to the list identified by `sibling_list_id`.
-    pub visit_integer:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `short` belonging to the list identified by `sibling_list_id`.
-    pub visit_short:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `byte` belonging to the list identified by `sibling_list_id`.
-    pub visit_byte:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `float` belonging to the list identified by `sibling_list_id`.
-    pub visit_float:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `double` belonging to the list identified by `sibling_list_id`.
-    pub visit_double:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `boolean` belonging to the list identified by `sibling_list_id`.
-    pub visit_boolean:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit `binary` belonging to the list identified by `sibling_list_id`.
-    pub visit_binary:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `date` belonging to the list identified by `sibling_list_id`.
-    pub visit_date:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `timestamp` belonging to the list identified by `sibling_list_id`.
-    pub visit_timestamp:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-
-    /// Visit a `timestamp` with no timezone belonging to the list identified by `sibling_list_id`.
-    pub visit_timestamp_ntz:
-        extern "C" fn(data: *mut c_void, sibling_list_id: usize, name: KernelStringSlice),
-}
-
-/// Visit the schema of the passed `SnapshotHandle`, using the provided `visitor`. See the
-/// documentation of [`EngineSchemaVisitor`] for a description of how this visitor works.
-///
-/// This method returns the id of the list allocated to hold the top level schema columns.
-///
-/// # Safety
-///
-/// Caller is responsible for passing a valid snapshot handle and schema visitor.
-#[no_mangle]
-pub unsafe extern "C" fn visit_schema(
-    snapshot: Handle<SharedSnapshot>,
-    visitor: &mut EngineSchemaVisitor,
-) -> usize {
-    let snapshot = unsafe { snapshot.as_ref() };
-    // Visit all the fields of a struct and return the list of children
-    fn visit_struct_fields(visitor: &EngineSchemaVisitor, s: &StructType) -> usize {
-        let child_list_id = (visitor.make_field_list)(visitor.data, s.fields.len());
-        for field in s.fields() {
-            visit_schema_item(field.data_type(), field.name(), visitor, child_list_id);
-        }
-        child_list_id
-    }
-
-    fn visit_array_item(visitor: &EngineSchemaVisitor, at: &ArrayType) -> usize {
-        let child_list_id = (visitor.make_field_list)(visitor.data, 1);
-        visit_schema_item(&at.element_type, "array_element", visitor, child_list_id);
-        child_list_id
-    }
-
-    fn visit_map_types(visitor: &EngineSchemaVisitor, mt: &MapType) -> usize {
-        let child_list_id = (visitor.make_field_list)(visitor.data, 2);
-        visit_schema_item(&mt.key_type, "map_key", visitor, child_list_id);
-        visit_schema_item(&mt.value_type, "map_value", visitor, child_list_id);
-        child_list_id
-    }
-
-    // Visit a struct field (recursively) and add the result to the list of siblings.
-    fn visit_schema_item(
-        data_type: &DataType,
-        name: &str,
-        visitor: &EngineSchemaVisitor,
-        sibling_list_id: usize,
-    ) {
-        macro_rules! call {
-            ( $visitor_fn:ident $(, $extra_args:expr) *) => {
-                (visitor.$visitor_fn)(visitor.data, sibling_list_id, name.into() $(, $extra_args) *)
-            };
-        }
-        match data_type {
-            DataType::Struct(st) => call!(visit_struct, visit_struct_fields(visitor, st)),
-            DataType::Map(mt) => {
-                call!(
-                    visit_map,
-                    mt.value_contains_null,
-                    visit_map_types(visitor, mt)
-                )
-            }
-            DataType::Array(at) => {
-                call!(visit_array, at.contains_null, visit_array_item(visitor, at))
-            }
-            DataType::Primitive(PrimitiveType::Decimal(precision, scale)) => {
-                call!(visit_decimal, *precision, *scale)
-            }
-            &DataType::STRING => call!(visit_string),
-            &DataType::LONG => call!(visit_long),
-            &DataType::INTEGER => call!(visit_integer),
-            &DataType::SHORT => call!(visit_short),
-            &DataType::BYTE => call!(visit_byte),
-            &DataType::FLOAT => call!(visit_float),
-            &DataType::DOUBLE => call!(visit_double),
-            &DataType::BOOLEAN => call!(visit_boolean),
-            &DataType::BINARY => call!(visit_binary),
-            &DataType::DATE => call!(visit_date),
-            &DataType::TIMESTAMP => call!(visit_timestamp),
-            &DataType::TIMESTAMP_NTZ => call!(visit_timestamp_ntz),
-        }
-    }
-
-    visit_struct_fields(visitor, snapshot.schema())
-}
-
-// TODO move expression visitors to separate module
-
 // A set that can identify its contents by address
 pub struct ReferenceSet<T> {
     map: std::collections::HashMap<usize, T>,
@@ -1046,241 +840,4 @@ impl<T> Default for ReferenceSet<T> {
             next_id: 1,
         }
     }
-}
-
-#[derive(Default)]
-pub struct KernelExpressionVisitorState {
-    // TODO: ReferenceSet<Box<dyn MetadataFilterFn>> instead?
-    inflight_expressions: ReferenceSet<Expression>,
-}
-impl KernelExpressionVisitorState {
-    fn new() -> Self {
-        Self {
-            inflight_expressions: Default::default(),
-        }
-    }
-}
-
-/// A predicate that can be used to skip data when scanning.
-///
-/// When invoking [`scan::scan`], The engine provides a pointer to the (engine's native) predicate,
-/// along with a visitor function that can be invoked to recursively visit the predicate. This
-/// engine state must be valid until the call to `scan::scan` returns. Inside that method, the
-/// kernel allocates visitor state, which becomes the second argument to the predicate visitor
-/// invocation along with the engine-provided predicate pointer. The visitor state is valid for the
-/// lifetime of the predicate visitor invocation. Thanks to this double indirection, engine and
-/// kernel each retain ownership of their respective objects, with no need to coordinate memory
-/// lifetimes with the other.
-#[repr(C)]
-pub struct EnginePredicate {
-    predicate: *mut c_void,
-    visitor:
-        extern "C" fn(predicate: *mut c_void, state: &mut KernelExpressionVisitorState) -> usize,
-}
-
-fn wrap_expression(state: &mut KernelExpressionVisitorState, expr: Expression) -> usize {
-    state.inflight_expressions.insert(expr)
-}
-
-fn unwrap_kernel_expression(
-    state: &mut KernelExpressionVisitorState,
-    exprid: usize,
-) -> Option<Expression> {
-    state.inflight_expressions.take(exprid)
-}
-
-fn visit_expression_binary(
-    state: &mut KernelExpressionVisitorState,
-    op: BinaryOperator,
-    a: usize,
-    b: usize,
-) -> usize {
-    let left = unwrap_kernel_expression(state, a).map(Box::new);
-    let right = unwrap_kernel_expression(state, b).map(Box::new);
-    match left.zip(right) {
-        Some((left, right)) => {
-            wrap_expression(state, Expression::BinaryOperation { op, left, right })
-        }
-        None => 0, // invalid child => invalid node
-    }
-}
-
-fn visit_expression_unary(
-    state: &mut KernelExpressionVisitorState,
-    op: UnaryOperator,
-    inner_expr: usize,
-) -> usize {
-    unwrap_kernel_expression(state, inner_expr).map_or(0, |expr| {
-        wrap_expression(state, Expression::unary(op, expr))
-    })
-}
-
-// The EngineIterator is not thread safe, not reentrant, not owned by callee, not freed by callee.
-#[no_mangle]
-pub extern "C" fn visit_expression_and(
-    state: &mut KernelExpressionVisitorState,
-    children: &mut EngineIterator,
-) -> usize {
-    let result = Expression::and_from(
-        children.flat_map(|child| unwrap_kernel_expression(state, child as usize)),
-    );
-    wrap_expression(state, result)
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_lt(
-    state: &mut KernelExpressionVisitorState,
-    a: usize,
-    b: usize,
-) -> usize {
-    visit_expression_binary(state, BinaryOperator::LessThan, a, b)
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_le(
-    state: &mut KernelExpressionVisitorState,
-    a: usize,
-    b: usize,
-) -> usize {
-    visit_expression_binary(state, BinaryOperator::LessThanOrEqual, a, b)
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_gt(
-    state: &mut KernelExpressionVisitorState,
-    a: usize,
-    b: usize,
-) -> usize {
-    visit_expression_binary(state, BinaryOperator::GreaterThan, a, b)
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_ge(
-    state: &mut KernelExpressionVisitorState,
-    a: usize,
-    b: usize,
-) -> usize {
-    visit_expression_binary(state, BinaryOperator::GreaterThanOrEqual, a, b)
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_eq(
-    state: &mut KernelExpressionVisitorState,
-    a: usize,
-    b: usize,
-) -> usize {
-    visit_expression_binary(state, BinaryOperator::Equal, a, b)
-}
-
-/// # Safety
-/// The string slice must be valid
-#[no_mangle]
-pub unsafe extern "C" fn visit_expression_column(
-    state: &mut KernelExpressionVisitorState,
-    name: KernelStringSlice,
-    allocate_error: AllocateErrorFn,
-) -> ExternResult<usize> {
-    let name = unsafe { String::try_from_slice(&name) };
-    visit_expression_column_impl(state, name).into_extern_result(&allocate_error)
-}
-fn visit_expression_column_impl(
-    state: &mut KernelExpressionVisitorState,
-    name: DeltaResult<String>,
-) -> DeltaResult<usize> {
-    Ok(wrap_expression(state, Expression::Column(name?)))
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_not(
-    state: &mut KernelExpressionVisitorState,
-    inner_expr: usize,
-) -> usize {
-    visit_expression_unary(state, UnaryOperator::Not, inner_expr)
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_is_null(
-    state: &mut KernelExpressionVisitorState,
-    inner_expr: usize,
-) -> usize {
-    visit_expression_unary(state, UnaryOperator::IsNull, inner_expr)
-}
-
-/// # Safety
-/// The string slice must be valid
-#[no_mangle]
-pub unsafe extern "C" fn visit_expression_literal_string(
-    state: &mut KernelExpressionVisitorState,
-    value: KernelStringSlice,
-    allocate_error: AllocateErrorFn,
-) -> ExternResult<usize> {
-    let value = unsafe { String::try_from_slice(&value) };
-    visit_expression_literal_string_impl(state, value).into_extern_result(&allocate_error)
-}
-fn visit_expression_literal_string_impl(
-    state: &mut KernelExpressionVisitorState,
-    value: DeltaResult<String>,
-) -> DeltaResult<usize> {
-    Ok(wrap_expression(
-        state,
-        Expression::Literal(Scalar::from(value?)),
-    ))
-}
-
-// We need to get parse.expand working to be able to macro everything below, see issue #255
-
-#[no_mangle]
-pub extern "C" fn visit_expression_literal_int(
-    state: &mut KernelExpressionVisitorState,
-    value: i32,
-) -> usize {
-    wrap_expression(state, Expression::literal(value))
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_literal_long(
-    state: &mut KernelExpressionVisitorState,
-    value: i64,
-) -> usize {
-    wrap_expression(state, Expression::literal(value))
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_literal_short(
-    state: &mut KernelExpressionVisitorState,
-    value: i16,
-) -> usize {
-    wrap_expression(state, Expression::literal(value))
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_literal_byte(
-    state: &mut KernelExpressionVisitorState,
-    value: i8,
-) -> usize {
-    wrap_expression(state, Expression::literal(value))
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_literal_float(
-    state: &mut KernelExpressionVisitorState,
-    value: f32,
-) -> usize {
-    wrap_expression(state, Expression::literal(value))
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_literal_double(
-    state: &mut KernelExpressionVisitorState,
-    value: f64,
-) -> usize {
-    wrap_expression(state, Expression::literal(value))
-}
-
-#[no_mangle]
-pub extern "C" fn visit_expression_literal_bool(
-    state: &mut KernelExpressionVisitorState,
-    value: bool,
-) -> usize {
-    wrap_expression(state, Expression::literal(value))
 }
